Refactor MapView components to follow standards
