+++ src/views/MapView/components/AddressPopup/index.js
@@ -1,0 +1,5 @@
+import { withStyles } from '@material-ui/core';
+import AddressPopup from './AddressPopup';
+import styles from '../../styles';
+
+export default withStyles(styles)(AddressPopup);
