+++ src/views/MapView/components/TransitStops/TransitStopInfo/index.js
@@ -1,0 +1,6 @@
+
+import { withStyles } from '@material-ui/core';
+import TransitStopInfo from './TransitStopInfo';
+import styles from '../../../styles';
+
+export default withStyles(styles)(TransitStopInfo);
