+++ src/views/MapView/components/TransitStops/index.js
@@ -1,0 +1,5 @@
+import { withStyles } from '@material-ui/core';
+import styles from '../../styles';
+import TransitStops from './TransitStops';
+
+export default withStyles(styles)(TransitStops);
