+++ src/views/MapView/components/UserMarker/index.js
@@ -1,0 +1,3 @@
+import UserMarker from './UserMarker';
+
+export default UserMarker;
